Corrigindo alguns erros com implementação da função de depósito
