+++ OpcoesContaBancaria.php
@@ -4,8 +4,9 @@
         <meta charset="utf-8">
         <title>Abertura de contas</title>
     </head>
-    <body> 
-        <form  method="post" action="MudancaMonetaria.php">
+    <body>
+
+        <form method="GET" action="MudancaMonetaria.php">
             <p><b>Escolha uma das opções a seguir:</b></p>
             <p>1-Para sacar dinheiro;</p>
             <p>2-Para depositar dinheiro</p>
@@ -13,7 +14,8 @@
             <p><input type="text" name="tipo"></p>
             <p>Digite o valor que deseja transferir/depositar/sacar</p>
             <p><input type="text" name="valor"></p>
-            <input type="submit" value="Enviar">
+            <input type="hidden" name="name" value="<?php echo $_GET['name'] ?>">
+            <input type="submit" value="Realizar">
         </form>
     </body>
 </html>
