+++ index.php
@@ -1,0 +1,14 @@
+<!DOCTYPE html> 
+<html lang="en" dir="ltr">
+    <head>
+        <meta charset="utf-8">
+        <title>Abertura de contas</title>
+    </head>
+    <body> 
+        <form  method="get" action="MudancaMonetaria.php">
+            <b>Digite seu nome para acessar sua conta:</b>:
+            <p><input type="text" name="name"></p>
+            <input type="submit" value="enviar">
+        </form>
+    </body>
+</html>
